Add initial code for to fix issue 2
The handler and the poller have been updated to add the changes
belonging to the toggle_study_status script.

select_study_task.psp now handles whether or not a user is an admin and
if the user is an admin, then the web interface will display an option
to make the study public or private depending on the current status.

toggle_study_status.psp will now fire up a job that will start the data
transfer over an ssh session.

The remaining work is plugging/testing this and adding an SQL procedure
to change the paths of the existing collated data.